requirements.txt
paho-mqtt
psycopg2
Adafruit_BMP
RPi.GPIO
adafruit-blinka
adafruit-circuitpython-dht
configparser
wheel
smbus